fix(#1273): deduplicate overlappingBoxes/stackedBoxes boolean fixtures
- Create BooleanTestFixtures.swift with shared overlappingBoxes() and stackedBoxes()
- Update Issue206BooleanTimeoutTests, Issue1067BooleanOutcomeTests, Issue832BooleanDelegationTests, Issue202BooleanOptionsTests to use shared fixtures
- All 19 boolean tests pass
- Gate scripts pass
